requirements.txt
beautifulsoup4
PyYaml
requests
json